fix(notifications): address remaining review issues
- Wrap delete action in $transaction for atomicity
- Fire mark-read in parallel with navigation instead of blocking
- Extract shared useNotifications hook, remove duplicate queryFn
- Add missing revalidatePath to markNotificationsReadByCardAction
- Clean up existing notifications before re-seeding
- Remove unused lib/notifications.ts helper
